Add import for time pacakge
Hi professor Hulten. I'm a student in CSEP 546
When you run the logistic regression model against the spam model you'll run into an error because there's no import for the time package that get's used  in the fit method. It was an easy fix so I thought I'd open a pr for it.